% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{asyncLog}
\alias{asyncLog}
\title{asyncLog}
\usage{
asyncLog(async, message, type = "default")
}
\arguments{
\item{async}{Whether the function is being used asynchronously}

\item{message}{The message to write to the logger}

\item{type}{One of "default", "info", "error", "warning"}
}
\description{
For internal use. Similar to writeLog but for use inside async
functions
}
\keyword{internal}
