% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_url}
\alias{check_url}
\title{check_url}
\usage{
check_url(url)
}
\arguments{
\item{url}{character. The URL to check}
}
\value{
An httr2 response if the URL is live or NULL if it is not
}
\description{
For internal use. Checks whether a URL is live
}
\keyword{internal}
