% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{close_loading_modal}
\alias{close_loading_modal}
\title{close_loading_modal}
\usage{
close_loading_modal(session = getDefaultReactiveDomain())
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}
}
\description{
For internal use. Close the modal once loading is complete
}
\keyword{internal}
