% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Create a skeleton app}
\usage{
init(
  path,
  name,
  include_map,
  include_table,
  include_code,
  common_objects,
  modules,
  author
)
}
\arguments{
\item{path}{character. The path to where the app should be created}

\item{name}{character. The name of the app which will be used as the package name}

\item{include_map}{logical. Whether to include a leaflet map}

\item{include_table}{logical. Whether to include a table tab}

\item{include_code}{logical. Whether to include a tab for viewing module code}

\item{common_objects}{character vector. Names of objects which will be shared between modules}

\item{modules}{A dataframe containing long and short names of components (tabs) and modules
in the order to be included and whether they should include mapping, save,
markdown and result functionality. The component and module columns are used to generate file
names for the modules. The long_component and long_module columns are used to generate UI and
so should be formatted appropriately.}

\item{author}{character. The name of the author(s)}
}
\description{
This function creates a skeleton app containing empty modules
}
\examples{
\dontrun{
modules <- data.frame(
"component" = c("data", "data", "plot", "plot"),
"long_component" = c("Load data", "Load data", "Plot data", "Plot data"),
"module" = c("user", "database", "histogram", "scatter"),
"long_module" = c("Upload your own data", "Query a database to obtain data",
"Plot the data as a histogram", "Plot the data as a scatterplot"),
"map" = c(TRUE, TRUE, FALSE, FALSE),
"result" = c(FALSE, FALSE, TRUE, TRUE),
"rmd" = c(TRUE, TRUE, TRUE, TRUE),
"save" = c(TRUE, TRUE, TRUE, TRUE))
common_objects = c("raster", "histogram", "scatter")
init(path = "~/Documents", name = "demo",
include_map = TRUE, include_table = TRUE, include_code = TRUE,
common_objects = common_objects, modules = modules,
author = "Simon E. H. Smart")
}
}
\author{
Simon E. H. Smart <simon.smart@cantab.net>
}
