% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_f.R
\name{plot_hist}
\alias{plot_hist}
\title{plot_hist}
\usage{
plot_hist(ras, bins)
}
\arguments{
\item{ras}{SpatRaster object}

\item{bins}{The number of breaks in the histogram}
}
\value{
a vector of values extracted from the raster
}
\description{
This function is called by the plot_hist module and extracts
values from a raster image.
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
