% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{plot_scatter}
\usage{
plot_scatter(ras, sample, axis)
}
\value{
a dataframe containing the axis values and the cell values
}
\description{
This function is called by the plot_scatter module and samples
 values from a raster
 @param ras SpatRaster object. Image to be analysed
 @param sample numeric. Number of points to sample
 @param axis character.
}
\author{
Simon Smart <simon.smart@cantab.net>
}
