% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{printVecAsis}
\alias{printVecAsis}
\title{printVecAsis}
\usage{
printVecAsis(x)
}
\arguments{
\item{x}{object to print}
}
\value{
A character string to reproduce the object
}
\description{
For internal use. Print objects as character string
}
\keyword{internal}
