% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_modules.R
\name{register_module}
\alias{register_module}
\title{Register a SMART module}
\usage{
register_module(config_file)
}
\arguments{
\item{config_file}{The path to a YAML file that contains the information about
one or more modules.}
}
\description{
Before running the SMART application with \code{run_smart()}, you can
register your own modules to be used in SMART.
}
\seealso{
\code{\link[SMART]{create_module}}
}
