% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{reset_data}
\alias{reset_data}
\title{reset_data}
\usage{
reset_data(common)
}
\arguments{
\item{common}{The common data structure}
}
\description{
For internal use. Clears the common structure of data, resets the map and any plots
}
\keyword{internal}
