% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_module.R
\name{run_module}
\alias{run_module}
\title{Run an individual \emph{shinyscholar} module for testing}
\usage{
run_module(module)
}
\arguments{
\item{module}{The name of the module to run}
}
\description{
This function runs the selected \emph{shinyscholar} module
}
\examples{
if(interactive()) {
run_module("select_query")
}
}
\author{
Simon E. H. Smart <simon.smart@cantab.net>
}
