% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_smart.R
\name{run_smart}
\alias{run_smart}
\title{Run \emph{SMART} Application}
\usage{
run_smart(launch.browser = TRUE, port = getOption("shiny.port"))
}
\arguments{
\item{launch.browser}{Whether or not to launch a new browser window.}

\item{port}{The port for the shiny server to listen on. Defaults to a
random available port.}
}
\description{
This function runs the \emph{SMART} application in the user's
default web browser.
}
\examples{
if(interactive()) {
run_smart()
}
}
\author{
Jamie Kass <jkass@gradcenter.cuny.edu>

Gonzalo E. Pinilla-Buitrago <gpinillabuitrago@gradcenter.cuny.edu>

Simon E. H. Smart <simon.smart@cantab.net>
}
