% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_query.R
\name{select_query}
\alias{select_query}
\title{select_query}
\usage{
select_query(poly, date, logger = NULL)
}
\arguments{
\item{poly}{matrix. Coordinates of area to load}

\item{date}{character. Date of image to load}
}
\value{
a SpatRaster object
}
\description{
This function is called by the select_query module and loads an
 FCover raster for the selected area.
}
\examples{
\dontrun{
 poly <- matrix(c(0, 0, 0.5, 0.5, 0, 52, 52.5, 52.5, 52, 52), ncol=2)
 colnames(poly) <- c("longitude", "latitude")
 date <- "2023-06-20"
 ras <- select_query(poly, date)
 }
}
\author{
Simon Smart <simon.smart@cantab.net>
}
