% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_user.R
\name{select_user}
\alias{select_user}
\title{select_user}
\usage{
select_user(ras_path)
}
\value{
a SpatRaster object
}
\description{
This function is called by the select_user module and loads a
 raster image.
 @param ras_path character. Path to file to be loaded
}
\author{
Simon Smart <simon.smart@cantab.net>
}
