% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{show_map}
\alias{show_map}
\title{show_map}
\usage{
show_map(parent_session)
}
\arguments{
\item{parent_session}{Session object of the main server function}
}
\value{
No return value, called for side effects
}
\description{
For internal use. Switches the view to the Map tab
}
\keyword{internal}
