% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{show_results}
\alias{show_results}
\title{show_results}
\usage{
show_results(parent_session)
}
\arguments{
\item{parent_session}{Session object of the main server function}
}
\description{
For internal use. Switches the view to the Results tab
}
\keyword{internal}
