% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{show_table}
\alias{show_table}
\title{show_table}
\usage{
show_table(parent_session)
}
\arguments{
\item{parent_session}{Session object of the main server function}
}
\description{
For internal use. Switches the view to the Table panel
}
\keyword{internal}
