% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tester.R
\name{test_module}
\alias{test_module}
\title{Run an individual \emph{SMART} module for testing}
\usage{
test_module(module)
}
\arguments{
\item{module}{The name of the module to run}
}
\description{
This function runs the selected \emph{SMART} module
}
\examples{
if(interactive()) {
test_module("select_query")
}
}
\author{
Simon E. H. Smart <simon.smart@cantab.net>
}
