% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{tidy_purl}
\alias{tidy_purl}
\title{tidy_purl}
\usage{
tidy_purl(params)
}
\arguments{
\item{params}{vector. Containing file object linking to file to be knitted
and a list containing any objects to be knitted into the file.}
}
\description{
Knits and purls an Rmd file into a format which can be written
to an R file.
}
\author{
Simon E. H. Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
\keyword{internal}
