% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wallace-package.R
\docType{package}
\name{wallace-package}
\alias{wallace-package}
\alias{wallace}
\title{\emph{SMART}: A modular platform for creating reproducible applications}
\description{
\emph{SMART} is a \code{shiny} app template for creating
  reproducible applications to academic standards. SMART is forked from 'wallace'
  and provides a template for producing applications that are interactive,
  reproducible, adaptable and built to high standards. It was created to ease
  the process of creating complex workflows in a single, streamlined GUI interface.
  In addition, executable session code  (R Markdown format) can
   be downloaded to share with others or use as supplementary information
   for scientific papers and reports. The application is run via the
   function \code{\link{run_smart}}.
}
\details{
Please see the official website (\url{https://wallaceecomod.github.io/}) for
  more details. If you have questions about the application, please participate
  in the \href{https://groups.google.com/forum/#!forum/wallaceecomod}{Google Group},
  or email the team directly: \email{wallaceEcoMod@gmail.com}.
}
