% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{asyncLog}
\alias{asyncLog}
\title{asyncLog}
\usage{
asyncLog(async, ..., type = "default")
}
\arguments{
\item{async}{Whether the function is being used asynchronously}

\item{...}{Messages to write to the logger}

\item{type}{One of \code{default}, \code{info}, \code{error}, \code{warning}}
}
\value{
No return value, called for side effects
}
\description{
For internal use. Similar to writeLog but for use inside async
functions
}
\keyword{internal}
