% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_suggests.R
\name{check_suggests}
\alias{check_suggests}
\title{Check suggests}
\usage{
check_suggests(testing = FALSE)
}
\arguments{
\item{testing.}{logical. For use in testing.}
}
\description{
Checks whether all the packages in Suggests are installed and stops execution
if not.
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
\keyword{internal}
