% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_modules.R
\name{create_module}
\alias{create_module}
\title{Create a shinyscholar module}
\usage{
create_module(
  id,
  dir,
  map = FALSE,
  result = FALSE,
  rmd = FALSE,
  save = FALSE,
  async = FALSE,
  init = FALSE
)
}
\arguments{
\item{id}{character. The id of the module.}

\item{dir}{character. A directory where the new module should be created.}

\item{map}{logical. Whether or not the module should support modifying the map.}

\item{result}{logical. Whether or not the module should support showing information in
the Result tab.}

\item{rmd}{logical. Whether or not the module should add Rmd code to the Session Code
download.}

\item{save}{logical. Whether or not the module has some custom data to save when the
user saves the current session.}

\item{async}{logical. Whether or not the module will operate asynchronously.}

\item{init}{logical. Whether or not the function is being used inside of the init function}
}
\description{
Create the template of a new shinyscholar module.
}
\seealso{
\code{\link[shinyscholar]{register_module}}
}
