% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_query_f.R
\name{get_nasa_token}
\alias{get_nasa_token}
\title{Fetch a token from the NASA Earthdata API}
\usage{
get_nasa_token(username, password)
}
\arguments{
\item{username}{character. NASA Earthdata username}

\item{password}{character. NASA Earthdata password}
}
\value{
A character string containing the token
}
\description{
Uses the Earthdata API to fetch a token using the user's
username and password
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
