% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{plot_hist}
\usage{
plot_hist(ras)
}
\value{
a vector of values extracted from the raster
}
\description{
This function is called by the plot_hist module and extracts
 values from a raster image.
 @param ras SpatRaster object
}
\author{
Simon Smart <simon.smart@cantab.net>
}
