% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_f.R
\name{plot_hist}
\alias{plot_hist}
\title{plot_hist}
\usage{
plot_hist(raster, bins, logger = NULL)
}
\arguments{
\item{raster}{SpatRaster object}

\item{bins}{The number of breaks in the histogram}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
a list of class histogram
}
\description{
This function is called by the plot_hist module and extracts
values from a raster image, returning a histogram of density
}
\examples{
if (check_suggests(example = TRUE)) {
  raster <- terra::rast(ncol = 8, nrow = 8)
  raster[] <- sapply(1:terra::ncell(raster), function(x){
    rnorm(1, ifelse(x \%\% 8 != 0, x \%\% 8, 8), 3)})
  histogram <- plot_hist(raster, bins = 10)
} else {
  message('reinstall with install.packages("shinyscholar", dependencies = TRUE)
  to run this example')
}
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
