% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_load.R
\name{save_and_load}
\alias{save_and_load}
\title{save_and_load}
\usage{
save_and_load(folder_path, module = NULL)
}
\arguments{
\item{folder_path}{character. Path to the parent directory containing the application}

\item{module}{character. (optional) Name of a single module to edit}
}
\description{
Adds lines to modules to save and load input values. By default
all the modules in the application are edited. Currently only input
functions from \code{{shiny}} and \code{shinyWidgets::materialSwitch} are included.
}
\author{
Simon E. H. Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
