% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_load.R
\name{save_and_load}
\alias{save_and_load}
\title{Adds lines to modules to save and load input values.}
\usage{
save_and_load(folder_path, module = NULL)
}
\arguments{
\item{folder_path}{character. Path to the parent directory containing the application}

\item{module}{character. (optional) Name of a single module to edit}
}
\description{
Converts 'shiny' \verb{*Input} functions to lines of code required
to store and reload the values when the app is saved or loaded. By default
all the modules in the application are edited. Currently only input
functions from 'shiny' and \code{shinyWidgets::materialSwitch} are supported.
}
\author{
Simon E. H. Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
