% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_async.R
\name{select_async}
\alias{select_async}
\title{select_async}
\usage{
select_async(poly, date, async = FALSE)
}
\arguments{
\item{poly}{matrix. Coordinates of area to load}

\item{date}{character. Date of image to load}

\item{async}{logical. Whether the function is being run asychronously}
}
\value{
A list containing:
 \item{raster}{a SpatRaster object when `async` is `FALSE` or a PackedSpatRaster when
`async` is `TRUE`}
 \item{message}{Information on the number of missing pixels}
}
\description{
This function is identical to select_query but can be run asychronously
}
\examples{
\dontrun{
 poly <- matrix(c(0, 0, 0.5, 0.5, 0, 52, 52.5, 52.5, 52, 52), ncol=2)
 colnames(poly) <- c("longitude", "latitude")
 date <- "2023-06-20"
 ras <- select_async(poly, date)
 }
}
\author{
Simon Smart <simon.smart@cantab.net>
}
