% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_query_f.R
\name{select_query}
\alias{select_query}
\title{select_query}
\usage{
select_query(poly, date, token, logger = NULL)
}
\arguments{
\item{poly}{matrix. Coordinates of area to load}

\item{date}{character.  Date of image to load in \code{YYYY-MM-DD} format.}

\item{token}{character. NASA Earthdata API token.
\href{https://urs.earthdata.nasa.gov/}{Click here}
to register and then
\href{https://wiki.earthdata.nasa.gov/pages/viewpage.action?pageId=204802786}{follow these instructions}
to obtain one. Alternatively supply your username and password to
\code{get_nasa_token()}}

\item{logger}{Stores all notification messages to be displayed in the Log
Window. Insert the logger reactive list here for running in
shiny, otherwise leave the default NULL}
}
\value{
a SpatRaster object
}
\description{
This function is called by the select_query module and loads an
FAPAR raster for the selected area.
}
\examples{
\dontrun{
 poly <- matrix(c(0.5, 0.5, 1, 1, 0.5, 52, 52.5, 52.5, 52, 52), ncol = 2)
 colnames(poly) <- c("longitude", "latitude")
 date <- "2023-06-20"
 token <- get_nasa_token(username = "<username>", password = "<password>")
 ras <- select_query(poly, date, token)
 }
}
\author{
Simon Smart \href{mailto:simon.smart@cantab.net}{simon.smart@cantab.net}
}
