% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_query.R
\name{select_query}
\alias{select_query}
\title{select_query}
\usage{
select_query(poly, date, logger = NULL)
}
\value{
a SpatRaster object
}
\description{
This function is called by the select_query module and loads an
 FCover raster for the selected area.
 @param extent SpatExtent object of the area
 @param date character. Date of image to load
 @examples
 \dontrun{
 extent <- c(43,50,-26,-12)
 date <- '2023-06-20'
 ras <- select_query(extent,date)
 }
}
\author{
Simon Smart <simon.smart@cantab.net>
}
