% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyscholar-package.R
\name{shinyscholar-package}
\alias{shinyscholar-package}
\alias{shinyscholar}
\title{\emph{shinyscholar}: A modular platform for creating reproducible applications}
\description{
\emph{shinyscholar} is a template for creating reproducible \code{shiny}
applications to academic standards. shinyscholar is forked from 'wallace'
and provides a template for producing applications that are interactive,
reproducible, adaptable and built to high standards. It was created to ease
the process of creating complex workflows in a single, streamlined GUI interface.
In addition, executable session code  (R Markdown format) can be
downloaded to share with others or use as supplementary information
for scientific papers and reports. An example application is run via the
function \code{\link{run_shinyscholar}} and requires
\code{install.packages("shinyscholar", dependencies = TRUE)}.
}
